refactor(cli)!: remove `common.py`
