+++ src/vectorcode/database/utils.py
@@ -3,8 +3,12 @@ import logging
 import os
 import socket
 import uuid
+from functools import cache
 
-from vectorcode.cli_utils import expand_path
+import chromadb
+from chromadb.utils import embedding_functions
+
+from vectorcode.cli_utils import Config, expand_path
 
 logger = logging.getLogger(name=__name__)
 
@@ -41,3 +45,27 @@ def get_collection_id(full_path: str) -> str:
         f"Hashing {plain_collection_name} as the collection name for {full_path}."
     )
     return collection_id
+
+
+@cache
+def get_embedding_function(configs: Config) -> chromadb.EmbeddingFunction:
+    try:
+        ef = getattr(embedding_functions, configs.embedding_function)(
+            **configs.embedding_params
+        )
+        if ef is None:  # pragma: nocover
+            raise AttributeError()
+        return ef
+    except AttributeError:
+        logger.warning(
+            f"Failed to use {configs.embedding_function}. Falling back to Sentence Transformer.",
+        )
+        return embedding_functions.SentenceTransformerEmbeddingFunction()  # type:ignore
+    except Exception as e:
+        e.add_note(
+            "\nFor errors caused by missing dependency, consult the documentation of pipx (or whatever package manager that you installed VectorCode with) for instructions to inject libraries into the virtual environment."
+        )
+        logger.error(
+            f"Failed to use {configs.embedding_function} with following error.",
+        )
+        raise
